refactor: capitalize page names
- consistent naming convention
