+++ src/App.jsx
@@ -2,8 +2,8 @@ import { Route, BrowserRouter as Router, Routes } from 'react-router-dom'
 
 import Home from '@/pages/Home'
 import CareersHub from '@/pages/CareersHub'
-import ConnectionsPage from '@/pages/connections'
-import MediaPage from '@/pages/media'
+import ConnectionsPage from '@/pages/Connections'
+import MediaPage from '@/pages/Media'
 import PreviousEvents from '@/pages/PreviousEvents'
 import PreviousEvent from '@/pages/PreviousEvent'
 import NotFound from '@/pages/NotFound'
